ATTDEF: more corrections to member types of dxf_attdef_struct.

+++ src/attdef.h
@@ -158,7 +158,7 @@ dxf_attdef_struct
                 /*!< Hard pointer ID / handle of PlotStyleName object.\n
                  * Group code = 390.\n
                  * \since Introduced in version R2009. */
-        long color_value;
+        int32_t color_value;
                 /*!< A 24-bit color value that should be dealt with in
                  * terms of bytes with values of 0 to 255.\n
                  * The lowest byte is the blue value, the middle byte is
@@ -178,7 +178,7 @@ dxf_attdef_struct
                  * class-level transparency data.\n
                  * Group code = 430.\n
                  * \since Introduced in version R2004. */
-        long transparency;
+        int32_t transparency;
                 /*!< Transparency value.\n
                  * The group code cannot be used by custom entities for
                  * their own data because the group code is reserved for
@@ -335,12 +335,12 @@ int16_t dxf_attdef_get_lineweight (DxfAttdef *attdef);
 DxfAttdef *dxf_attdef_set_lineweight (DxfAttdef *attdef, int16_t lineweight);
 char *dxf_attdef_get_plot_style_name (DxfAttdef *attdef);
 DxfAttdef *dxf_attdef_set_plot_style_name (DxfAttdef *attdef, char *plot_style_name);
-long dxf_attdef_get_color_value (DxfAttdef *attdef);
-DxfAttdef *dxf_attdef_set_color_value (DxfAttdef *attdef, long color_value);
+int32_t dxf_attdef_get_color_value (DxfAttdef *attdef);
+DxfAttdef *dxf_attdef_set_color_value (DxfAttdef *attdef, int32_t color_value);
 char *dxf_attdef_get_color_name (DxfAttdef *attdef);
 DxfAttdef *dxf_attdef_set_color_name (DxfAttdef *attdef, char *color_name);
-long dxf_attdef_get_transparency (DxfAttdef *attdef);
-DxfAttdef *dxf_attdef_set_transparency (DxfAttdef *attdef, long transparency);
+int32_t dxf_attdef_get_transparency (DxfAttdef *attdef);
+DxfAttdef *dxf_attdef_set_transparency (DxfAttdef *attdef, int32_t transparency);
 char *dxf_attdef_get_default_value (DxfAttdef *attdef);
 DxfAttdef *dxf_attdef_set_default_value (DxfAttdef *attdef, char *default_value);
 char *dxf_attdef_get_tag_value (DxfAttdef *attdef);
